Move PendingPrediction from runner.go to types.go
Move PendingPrediction type and methods to types.go since it's used by both
Manager and Runner components. This improves code organization by keeping
shared types in their designated location.

Changes:
- Move PendingPrediction struct definition to types.go
- Move all PendingPrediction methods (safeSend, safeClose, sendWebhook, sendWebhookSync)
- Add comprehensive tests for PendingPrediction methods in types_test.go
- Update concurrent operation tests to use new sync.WaitGroup.Go

No functional changes - purely organizational refactoring.
